requirements.txt
Flask==2.0.1
google-cloud-ndb
